+++ scripts/travis/install_deps_osx.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Install dependencies for Mac OSX
+echo "Running on Mac OSX."
+
+echo "Mac OSX version not implemented for Travis Build testing."
+exit 99