+++ scripts/ubuntu/install_caffe_and_openpose_JetsonTX2_JetPack3.1.sh
@@ -3,7 +3,7 @@
 
 
 echo "------------------------- Installing Caffe and OpenPose -------------------------"
-echo "NOTE: This script assumes that CUDA and cuDNN are already installed on your machine. Otherwise, it might fail."
+echo "NOTE: This script assumes that just flashed JetPack 3.1 : Ubuntu 16, CUDA 8, cuDNN 6 and OpenCV are already installed on your machine. Otherwise, it might fail."
 
 
 
@@ -36,12 +36,12 @@ function executeShInItsFolder {
 
 
 git submodule update --init --recursive
-executeShInItsFolder "install_caffe_if_cuda8.sh" "./3rdparty/caffe" "../.."
+executeShInItsFolder "install_caffe_JetsonTX2_JetPack3.1.sh" "./3rdparty/caffe" "../.."
 exitIfError
 
 
 
-executeShInItsFolder "install_openpose_if_cuda8.sh" "./3rdparty/ubuntu_deprecated/" "./"
+executeShInItsFolder "install_openpose_JetsonTX2_JetPack3.1.sh" "./scripts/ubuntu/" "./"
 exitIfError