+++ scripts/ubuntu/install_openpose_JetsonTX2_JetPack3.1.sh
@@ -3,9 +3,7 @@
 
 
 echo "------------------------- Installing OpenPose -------------------------"
-echo "NOTE: This script assumes that CUDA and cuDNN are already installed on your machine. Otherwise, it might fail."
-
-
+echo "NOTE: This script assumes that just flashed JetPack 3.1 : Ubuntu 16, CUDA 8, cuDNN 6 and OpenCV are already installed on your machine. Otherwise, it might fail."
 
 function exitIfError {
     if [[ $? -ne 0 ]] ; then
@@ -27,7 +25,7 @@ function executeShInItsFolder {
     exitIfError
     sudo chmod +x $1
     exitIfError
-    ./$1
+    bash ./$1
     exitIfError
     cd $3
     exitIfError
@@ -36,27 +34,8 @@ function executeShInItsFolder {
 
 
 echo "------------------------- Checking Ubuntu Version -------------------------"
-ubuntu_version="$(lsb_release -r)"
-echo "Ubuntu $ubuntu_version"
-if [[ $ubuntu_version == *"14."* ]]; then
-    ubuntu_le_14=true
-elif [[ $ubuntu_version == *"16."* || $ubuntu_version == *"15."* || $ubuntu_version == *"17."* || $ubuntu_version == *"18."* ]]; then
-    ubuntu_le_14=false
-else
-    echo "Ubuntu release older than version 14. This installation script might fail."
-    ubuntu_le_14=true
-fi
-exitIfError
+# If you respected the installation assumptions, nothing to do. 
 echo "------------------------- Ubuntu Version Checked -------------------------"
-echo ""
-
-
-
-echo "------------------------- Checking Number of Processors -------------------------"
-NUM_CORES=$(grep -c ^processor /proc/cpuinfo 2>/dev/null || sysctl -n hw.ncpu)
-echo "$NUM_CORES cores"
-exitIfError
-echo "------------------------- Number of Processors Checked -------------------------"
 echo ""
 
 
@@ -65,14 +44,10 @@ echo "------------------------- Compiling OpenPose -------------------------"
 # Go back to main folder
 cd ../..
 # Copy Makefile & Makefile.config
-cp 3rdparty/ubuntu/Makefile.example Makefile
-if [[ $ubuntu_le_14 == true ]]; then
-    cp 3rdparty/ubuntu_deprecated/Makefile.config.Ubuntu14_cuda8.example Makefile.config
-else
-    cp 3rdparty/ubuntu_deprecated/Makefile.config.Ubuntu16_cuda8.example Makefile.config
-fi
+cp scripts/ubuntu/Makefile.example Makefile
+cp scripts/ubuntu/Makefile.config.Ubuntu16_cuda8_JetsonTX2 Makefile.config
 # Compile OpenPose
-make all -j$NUM_CORES
+make all -j`nproc`
 exitIfError
 echo "------------------------- OpenPose Compiled -------------------------"
 echo ""