+++ scripts/ubuntu/install_openpose_JetsonTX2_JetPack3.3.sh
@@ -25,7 +25,7 @@ function executeShInItsFolder {
     exitIfError
     sudo chmod +x $1
     exitIfError
-    ./$1
+    bash ./$1
     exitIfError
     cd $3
     exitIfError
@@ -40,23 +40,14 @@ echo ""
 
 
 
-echo "------------------------- Checking Number of Processors -------------------------"
-NUM_CORES=$(grep -c ^processor /proc/cpuinfo 2>/dev/null || sysctl -n hw.ncpu)
-echo "$NUM_CORES cores"
-exitIfError
-echo "------------------------- Number of Processors Checked -------------------------"
-echo ""
-
-
-
 echo "------------------------- Compiling OpenPose -------------------------"
 # Go back to main folder
 cd ../..
 # Copy Makefile & Makefile.config
-cp 3rdparty/ubuntu/Makefile.example_openpose Makefile
-cp 3rdparty/ubuntu/Makefile.config.Ubuntu16_cuda9_JetsonTX2_openpose Makefile.config
+cp scripts/ubuntu/Makefile.example_openpose Makefile
+cp scripts/ubuntu/Makefile.config.Ubuntu16_cuda9_JetsonTX2_openpose Makefile.config
 # Compile OpenPose
-make all -j$NUM_CORES
+make all -j`nproc`
 exitIfError
 echo "------------------------- OpenPose Compiled -------------------------"
 echo ""